requirements.txt
pandas>=2.0.0
matplotlib>=3.8.0
Pillow>=10.0.0
requests>=2.31.0
nfl-data-py>=0.3.0
numpy>=1.24.0
adjustText>=0.8.1
scipy>=1.11.0
rembg>=2.0.50
pyinstaller>=6.0.0
python-dotenv>=1.0.0
